+++ src/main/java/com/caoccao/javet/values/reference/typed/V8ValueTypedArray.java
@@ -1,0 +1,51 @@
+/*
+ *   Copyright (c) 2021. caoccao.com Sam Cao
+ *   All rights reserved.
+ *
+ *   Licensed under the Apache License, Version 2.0 (the "License");
+ *   you may not use this file except in compliance with the License.
+ *   You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *   Unless required by applicable law or agreed to in writing, software
+ *   distributed under the License is distributed on an "AS IS" BASIS,
+ *   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *   See the License for the specific language governing permissions and
+ *   limitations under the License.
+ */
+
+package com.caoccao.javet.values.reference.typed;
+
+import com.caoccao.javet.exceptions.JavetException;
+import com.caoccao.javet.values.reference.V8ValueArrayBuffer;
+import com.caoccao.javet.values.reference.V8ValueObject;
+
+public abstract class V8ValueTypedArray extends V8ValueObject implements IV8ValueTypedArray {
+
+    public static final String PROPERTY_BYTE_LENGTH = "byteLength";
+    public static final String PROPERTY_BUFFER = "buffer";
+
+    public V8ValueTypedArray(long handle) {
+        super(handle);
+    }
+
+    @Override
+    public V8ValueArrayBuffer getArrayBuffer() throws JavetException {
+        return get(PROPERTY_BUFFER);
+    }
+
+    @Override
+    public int getByteLength() throws JavetException {
+        return getInteger(PROPERTY_BYTE_LENGTH);
+    }
+
+    @Override
+    public int getLength() throws JavetException {
+        checkV8Runtime();
+        return v8Runtime.getLength(this);
+    }
+
+    @Override
+    public abstract int getSizeInBytes();
+}