+++ docs/release_notes.rst
@@ -8,7 +8,8 @@ Release Notes
 * Added ``setFunction(String functionName, String codeString)`` to ``IV8ValueObject``
 * Added ``equals()`` and ``strictEquals()`` and ``sameValue()`` to ``IV8Value``
 * Added ``getIdentityHash()`` to ``IV8ValueReference``
-* Added ``isDead()``, ``isInUse()`` and ``terminateExecution()`` to ``V8Runtime``
+* Added ``isDead()``, ``isInUse()``, ``callAsConstructor()`` and ``terminateExecution()`` to ``V8Runtime``
+* Added V8 typed array
 * Added ``IJavetEngineGuard``
 
 0.7.1